Control.CanAquire

Description

Documentation

runAcquire :: forall t a b. Traversable t => (forall s. CanAquire (I s a) a => t (I s a) -> b) -> t a -> b

Run a computation on something that can aquire i's.

class HasIndex i Int => CanAquire i a where

Methods

aquire :: i -> a

A value of type i can obtain something of type 'a'

class HasIndex t i | t -> i where

Methods

indexOf :: t -> i

Types that have an instance of this class can act as indices.

replaceByIndex :: forall t a. Traversable t => t a -> (Vector a, t Int)

Replaces every element by an index. Returns the new traversable containing only these indices, as well as a vector with the values. (such that indexing in this value gives the original value).

labelWithIndex :: Traversable t => t a -> (t (Int, a), Int)

Label each element with its index. Returns the new collection as well as its size.

data I (s :: *) a

A type that can act as an Index.