Control.Monad.State.Persistent

Description

Documentation

data PersistentStateT s m a

A State monad that can store earlier versions of the state.

type PersistentState s = PersistentStateT s Identity

A State monad that can store earlier versions of the state.

store :: Monad m => PersistentStateT s m ()

Create a snapshot of the current state and add it to the list of states that we store.

runPersistentStateT :: Functor m => PersistentStateT s m a -> s -> m (a, s, [s])

Run a persistentStateT, returns a triplet with the value, the last state and a list of all states (including the last one) in chronological order

runPersistentState :: PersistentState s a -> s -> (a, s, [s])

Run a persistentStateT, returns a triplet with the value, the last state and a list of all states (including the last one) in chronological order