Data.Ext

Description

A pair-like data type to represent a core type that has extra information as well.

Documentation

data core :+ extra infixr 1

Our Ext type that represents the core datatype core extended with extra information of type extra.

Constructors

core :+ extra infixr 1

_core :: (core :+ extra) -> core

_extra :: (core :+ extra) -> extra

core :: Lens (core :+ extra) (core' :+ extra) core core'

extra :: Lens (core :+ extra) (core :+ extra') extra extra'

ext :: a -> a :+ ()