Data.IndexedDoublyLinkedList

Documentation

data DLList s a

Doubly linked list implemented by a mutable vector. So actually this data type can represent a collection of Linked Lists that can efficiently be concatenated and split.

Supports O(1) indexing, and O(1) insertions, deletions

Constructors

DLList

Fields

values :: !(Vector a)

llist :: !(MVector s Cell)

data Cell

Cells in the Linked List

Constructors

Cell

Fields

prev :: Maybe Index

next :: Maybe Index

emptyCell :: Cell

data DLListMonad s b a

Monad in which we can use the IndexedDoublyLinkedList.

runDLListMonad :: Vector b -> (forall s. DLListMonad s b a) -> a

Runs a DLList Computation, starting with singleton values, crated from the input vector.

type Index = Int

singletons :: (PrimMonad m, s ~ PrimState m) => Vector b -> m (DLList s b)

Constructs a new DoublyLinkedList. Every element is its own singleton list

writeList :: NonEmpty Index -> DLListMonad s b ()

Sets the DoublyLinkedList to the given List.

Indices that do not occur in the list are not touched.

valueAt :: Index -> DLListMonad s b b

Gets the value at Index i

getNext :: Index -> DLListMonad s b (Maybe Index)

Next element in the List

getPrev :: Index -> DLListMonad s b (Maybe Index)

Previous Element in the List

toListFrom :: Index -> DLListMonad s b (NonEmpty Index)

Computes a maximal length list starting from the Given index

running time: \(O(k)\), where \(k\) is the length of the output list

toListFromR :: Index -> DLListMonad s b (NonEmpty Index)

Computes a maximal length list by walking backwards in the DoublyLinkedList, starting from the Given index

running time: \(O(k)\), where \(k\) is the length of the output list

toListContains :: Index -> DLListMonad s b (NonEmpty Index)

Computes a maximal length list that contains the element i.

running time: \(O(k)\), where \(k\) is the length of the output list

toListFromK :: Index -> Int -> DLListMonad s b (NonEmpty Index)

Takes the current element and its k next's

toListFromRK :: Index -> Int -> DLListMonad s b (NonEmpty Index)

insertAfter :: Index -> Index -> DLListMonad s b ()

Inserts the second argument after the first one into the linked list

insertBefore :: Index -> Index -> DLListMonad s b ()

Inserts the second argument before the first one into the linked list

delete :: Index -> DLListMonad s b (Maybe Index, Maybe Index)

Deletes the element from the linked list. This element thus essentially becomes a singleton list. Returns the pair of indices that now have become neighbours (i.e. the predecessor and successor of j just before we deleted j).

dump :: DLListMonad s a (Vector a, Vector Cell)

For debugging purposes, dump the values and the cells