Data.IndexedDoublyLinkedList.Bare

Description

Documentation

newtype IDLList s

Doubly linked list implemented by a mutable vector. So actually this data type can represent a collection of Linked Lists that can efficiently be concatenated and split.

Supports O(1) indexing, and O(1) insertions, deletions

Constructors

IDLList

Fields

llist :: MVector s Cell

data Cell

Cells in the Linked List

Constructors

Cell

Fields

prev :: !(Maybe Index)

next :: !(Maybe Index)

emptyCell :: Cell

Empty cell with no next or prev cells.

data IDLListMonad s a

Monad in which we can use the IndexedDoublyLinkedList.

runIDLListMonad :: Int -> (forall s. IDLListMonad s a) -> a

Runs a DLList Computation, starting with n singleton values

type Index = Int

Cell indices. Must be non-negative.

singletons :: (PrimMonad m, s ~ PrimState m) => Int -> m (IDLList s)

Constructs a new DoublyLinkedList, of size at most n

writeList :: NonEmpty Index -> IDLListMonad s ()

Sets the DoublyLinkedList to the given List.

Indices that do not occur in the list are not touched.

getNext :: Index -> IDLListMonad s (Maybe Index)

Next element in the List

getPrev :: Index -> IDLListMonad s (Maybe Index)

Previous Element in the List

toListFrom :: Index -> IDLListMonad s (NonEmpty Index)

Computes a maximal length list starting from the Given index

running time: \(O(k)\), where \(k\) is the length of the output list

toListFromR :: Index -> IDLListMonad s (NonEmpty Index)

Computes a maximal length list by walking backwards in the DoublyLinkedList, starting from the Given index

running time: \(O(k)\), where \(k\) is the length of the output list

toListContains :: Index -> IDLListMonad s (NonEmpty Index)

Computes a maximal length list that contains the element i.

running time: \(O(k)\), where \(k\) is the length of the output list

toListFromK :: Index -> Int -> IDLListMonad s (NonEmpty Index)

Takes the current element and its k next's

toListFromRK :: Index -> Int -> IDLListMonad s (NonEmpty Index)

Takes the current element and its k prev's

insertAfter :: Index -> Index -> IDLListMonad s ()

Inserts the second argument after the first one into the linked list

insertBefore :: Index -> Index -> IDLListMonad s ()

Inserts the second argument before the first one into the linked list

delete :: Index -> IDLListMonad s ()

Deletes the element from the linked list. This element thus essentially becomes a singleton list.

dump :: IDLListMonad s (Vector Cell)

For debugging purposes, dump the values and the cells