Data.Intersection

Description

Defines a data type for representing intersections. Mostly useful for the more geometric types.

Documentation

data NoIntersection

A simple data type expressing that there are no intersections

Constructors

NoIntersection

type Intersection g h = CoRec Identity (IntersectionOf g h)

The result of interesecting two geometries is a CoRec,

type family IntersectionOf g h :: [*]

The type family specifying the list of possible result types of an intersection.

coRec :: a ∈ as => a -> CoRec Identity as

Helper to produce a corec

class IsIntersectableWith g h where

Minimal complete definition

intersect, nonEmptyIntersection

Methods

intersect :: g -> h -> Intersection g h

intersects :: g -> h -> Bool

g intersects h = The intersection of g and h is non-empty.

The default implementation computes the intersection of g and h, and uses nonEmptyIntersection to determine if the intersection is non-empty.

nonEmptyIntersection :: proxy g -> proxy h -> Intersection g h -> Bool

Helper to implement intersects.

default nonEmptyIntersection :: (NoIntersection ∈ IntersectionOf g h, RecApplicative (IntersectionOf g h)) => proxy g -> proxy h -> Intersection g h -> Bool

type AlwaysTrueIntersection g h = RecApplicative (IntersectionOf g h)

When using IntersectionOf we may need some constraints that are always true anyway.

defaultNonEmptyIntersection :: forall g h proxy. (NoIntersection ∈ IntersectionOf g h, RecApplicative (IntersectionOf g h)) => proxy g -> proxy h -> Intersection g h -> Bool

Returns True iff the result is *not* a NoIntersection