Data.List.Set

Description

Documentation

data Set a

A Set of a's, implemented using a simple list. The only advantage of this implementation over Set from containers is that most operations require only 'Eq a' rather than 'Ord a'.

singleton :: a -> Set a

Creates a singleton set.

insert :: Eq a => a -> Set a -> Set a

\(O(n)\) Inserts an element in the set

delete :: Eq a => a -> Set a -> Set a

\(O(n)\) Deletes an element from the set

union :: Eq a => Set a -> Set a -> Set a

\(O(n^2)\) Computes the union of two sets

intersection :: Eq a => Set a -> Set a -> Set a

\(O(n^2)\) Computes the intersection of two sets

difference :: Eq a => Set a -> Set a -> Set a

\(O(n^2)\) Computes the difference of two sets

fromList :: Eq a => [a] -> Set a

\( O(n^2) \) Create a set from a finite list of elements.

insertAll :: Eq a => [a] -> Set a -> Set a

\( O(n^2) \) Insert an element in a set.