Data.List.Zipper

Description

Documentation

data Zipper a

Simple Zipper for Lists.

Constructors

Zipper [a] [a]

fromList :: [a] -> Zipper a

Construct a Zipper from a list

running time: \(O(1)\)

goNext :: Zipper a -> Maybe (Zipper a)

Go to the Next Element

running time: \(O(1)\)

goPrev :: Zipper a -> Maybe (Zipper a)

Go to the previous Element

running time: \(O(1)\)

allNexts :: Zipper a -> [Zipper a]

Computes all nexts, even one that has no elements initially or at the end.

>>> mapM_ print $ allNexts $ fromList [1..5]
Zipper [] [1,2,3,4,5]
Zipper [1] [2,3,4,5]
Zipper [2,1] [3,4,5]
Zipper [3,2,1] [4,5]
Zipper [4,3,2,1] [5]
Zipper [5,4,3,2,1] []

extractNext :: Zipper a -> Maybe (a, Zipper a)

Returns the next element, and the zipper without it

dropNext :: Zipper a -> Maybe (Zipper a)

Drops the next element in the zipper.

running time: \(O(1)\)

allNonEmptyNexts :: Zipper a -> [Zipper a]

Computes all list that still have next elements.

>>> mapM_ print $ allNonEmptyNexts $ fromList [1..5]
Zipper [] [1,2,3,4,5]
Zipper [1] [2,3,4,5]
Zipper [2,1] [3,4,5]
Zipper [3,2,1] [4,5]
Zipper [4,3,2,1] [5]