Data.OrdSeq

Description

Documentation

data OrdSeq a

Sequence of ordered elements.

type Compare a = a -> a -> Ordering

Signature for functions that give the ordering of two values.

insertBy :: Compare a -> a -> OrdSeq a -> OrdSeq a

Insert into a monotone OrdSeq.

pre: the comparator maintains monotonicity

\(O(\log n)\)

insert :: Ord a => a -> OrdSeq a -> OrdSeq a

Insert into a sorted OrdSeq

\(O(\log n)\)

splitBy :: Compare a -> a -> OrdSeq a -> (OrdSeq a, OrdSeq a, OrdSeq a)

\(O(\log n)\)

splitOn :: Ord b => (a -> b) -> b -> OrdSeq a -> (OrdSeq a, OrdSeq a, OrdSeq a)

Given a monotonic function f that maps a to b, split the sequence s depending on the b values. I.e. the result (l,m,r) is such that

all (< x) . fmap f $ l

all (== x) . fmap f $ m

all (> x) . fmap f $ r

>>> splitOn id 3 $ fromAscList [1..5]
(fromAscList [1,2],fromAscList [3],fromAscList [4,5])
>>> splitOn fst 2 $ fromAscList [(0,"-"),(1,"A"),(2,"B"),(2,"C"),(3,"D"),(4,"E")]
(fromAscList [(0,"-"),(1,"A")],fromAscList [(2,"B"),(2,"C")],fromAscList [(3,"D"),(4,"E")])

splitMonotonic :: (a -> Bool) -> OrdSeq a -> (OrdSeq a, OrdSeq a)

Given a monotonic predicate p, splits the sequence s into two sequences (as,bs) such that all (not p) as and all p bs

\(O(\log n)\)

deleteAll :: Ord a => a -> OrdSeq a -> OrdSeq a

Deletes all elements from the OrdDeq

\(O(n\log n)\)

deleteAllBy :: Compare a -> a -> OrdSeq a -> OrdSeq a

\( O(\log n) \). Delete all elements that compare as equal to x.

fromListBy :: Compare a -> [a] -> OrdSeq a

inserts all eleements in order \(O(n\log n)\)

fromListByOrd :: Ord a => [a] -> OrdSeq a

inserts all eleements in order \(O(n\log n)\)

fromAscList :: [a] -> OrdSeq a

\( O(n) \)

lookupBy :: Compare a -> a -> OrdSeq a -> Maybe a

\(O(\log n)\)

memberBy :: Compare a -> a -> OrdSeq a -> Bool

\(O(\log n)\). Queries for the existance of any elements that compare as equal to x.

mapMonotonic :: (a -> b) -> OrdSeq a -> OrdSeq b

\( O(n) \) Fmap, assumes the order does not change

viewl :: OrdSeq a -> ViewL OrdSeq a

\(O(1)\) Gets the first element from the sequence

viewr :: OrdSeq a -> ViewR OrdSeq a

\(O(1)\) Gets the last element from the sequence

minView :: OrdSeq a -> Maybe (a, OrdSeq a)

\(O(1)\)

lookupMin :: OrdSeq a -> Maybe a

\(O(1)\)

maxView :: OrdSeq a -> Maybe (a, OrdSeq a)

\(O(1)\)

lookupMax :: OrdSeq a -> Maybe a

\(O(1)\)