Data.Permutation

Description

Data type for representing a Permutation

Documentation

type Orbit a = Vector a

Orbits (Cycles) are represented by vectors

data Permutation a

Cyclic representation of a permutation.

Constructors

Permutation

Fields

_orbits :: Vector (Orbit a)

_indexes :: Vector (Int, Int)

idxes (fromEnum a) = (i,j) implies that a is the j^th item in the i^th orbit

orbits :: forall a a. Lens (Permutation a) (Permutation a) (Vector (Orbit a)) (Vector (Orbit a))

indexes :: forall a. Lens' (Permutation a) (Vector (Int, Int))

elems :: Permutation a -> Vector a

size :: Permutation a -> Int

cycleOf :: Enum a => Permutation a -> a -> Orbit a

The cycle containing a given item

next :: Vector v a => v a -> Int -> a

Next item in a cyclic permutation

previous :: Vector v a => v a -> Int -> a

Previous item in a cyclic permutation

lookupIdx :: Enum a => Permutation a -> a -> (Int, Int)

Lookup the indices of an element, i.e. in which orbit the item is, and the index within the orbit.

runnign time: \(O(1)\)

apply :: Enum a => Permutation a -> a -> a

Apply the permutation, i.e. consider the permutation as a function.

orbitFrom :: Eq a => a -> (a -> a) -> [a]

Find the cycle in the permutation starting at element s

cycleRep :: (Vector v a, Enum a, Eq a) => v a -> (a -> a) -> Permutation a

Given a vector with items in the permutation, and a permutation (by its functional representation) construct the cyclic representation of the permutation.

toCycleRep :: Enum a => Int -> [[a]] -> Permutation a

Given the size n, and a list of Cycles, turns the cycles into a cyclic representation of the Permutation.

genIndexes :: Enum a => Int -> [[a]] -> Vector (Int, Int)