Data.PlanarGraph.AdjRep

Description

Data types that to represent a planar graph as Adjacency Lists. The main purpose is to help encodedecode a PlanarGraph as a JSONYAML file.

Documentation

data Gr v f

Data type representing the graph in its JSON/Yaml format

Constructors

Gr

Fields

adjacencies :: [v]

faces :: [f]

data Vtx v e

A vertex, represented by an id, its adjacencies, and its data.

Constructors

Vtx

Fields

id :: !Int

adj :: [(Int, e)]

adjacent vertices + data on the edge. Some functions, like fromAdjRep may assume that the adjacencies are given in counterclockwise order around the vertices. This is not (yet) enforced by the data type.

vData :: !v

data Face f

Faces

Constructors

Face

Fields

incidentEdge :: (Int, Int)

an edge (u,v) s.t. the face is right from (u,v)

fData :: !f