Data.PlanarGraph.Dart

Description

Data type for representing Darts (edges) in a planar graph.

Documentation

>>> :{
let dart i s = Dart (Arc i) (read s)
:}

newtype Arc s

An Arc is a directed edge in a planar graph. The type s is used to tie this arc to a particular graph.

Constructors

Arc

Fields

_unArc :: Int

data Direction

Darts have a direction which is either Positive or Negative (shown as +1 or -1, respectively).

Constructors

Negative

Positive

rev :: Direction -> Direction

Reverse the direcion

data Dart s

A dart represents a bi-directed edge. I.e. a dart has a direction, however the dart of the oposite direction is always present in the planar graph as well.

Constructors

Dart

Fields

_arc :: !(Arc s)

_direction :: !Direction

direction :: forall s. Lens' (Dart s) Direction

arc :: forall s s. Lens (Dart s) (Dart s) (Arc s) (Arc s)

twin :: Dart s -> Dart s

Get the twin of this dart (edge)

>>> twin (dart 0 "+1")
Dart (Arc 0) -1
>>> twin (dart 0 "-1")
Dart (Arc 0) +1

isPositive :: Dart s -> Bool

test if a dart is Positive

allDarts :: [Dart s]

Enumerates all darts such that allDarts !! i = d = i == fromEnum d