Data.Range

Description

Data type for representing Generic Ranges (Intervals) and functions that work with them.

Documentation

data EndPoint a

Endpoints of a range may either be open or closed.

Constructors

Open !a

Closed !a

isOpen :: EndPoint a -> Bool

True iff EndPoint is open.

isClosed :: EndPoint a -> Bool

True iff EndPoint is closed.

unEndPoint :: Lens (EndPoint a) (EndPoint b) a b

Access lens for EndPoint value regardless of whether it is open or closed.

>>> Open 5 ^. unEndPoint
5
>>> Closed 10 ^. unEndPoint
10
>>> Open 4 & unEndPoint .~ 0
Open 0

data Range a

Data type for representing ranges.

Constructors

Range

Fields

_lower :: !(EndPoint a)

_upper :: !(EndPoint a)

Bundled Patterns

pattern OpenRange :: a -> a -> Range a

pattern ClosedRange :: a -> a -> Range a

pattern Range' :: a -> a -> Range a
A range from l to u, ignoring/forgetting the type of the endpoints

prettyShow :: Show a => Range a -> String

Helper function to show a range in mathematical notation.

>>> prettyShow $ OpenRange 0 2
"(0,2)"
>>> prettyShow $ ClosedRange 0 2
"[0,2]"
>>> prettyShow $ Range (Open 0) (Closed 5)
"(0,5]"

lower :: Lens' (Range a) (EndPoint a)

Lens access for the lower part of a range.

upper :: Lens' (Range a) (EndPoint a)

Lens access for the upper part of a range.

inRange :: Ord a => a -> Range a -> Bool

Test if a value lies in a range.

>>> 1 `inRange` (OpenRange 0 2)
True
>>> 1 `inRange` (OpenRange 0 1)
False
>>> 1 `inRange` (ClosedRange 0 1)
True
>>> 1 `inRange` (ClosedRange 1 1)
True
>>> 10 `inRange` (OpenRange 1 10)
False
>>> 10 `inRange` (ClosedRange 0 1)
False

This one is kind of weird

>>> 0 `inRange` Range (Closed 0) (Open 0)
False

width :: Num r => Range r -> r

Get the width of the interval

>>> width $ ClosedRange 1 10
9
>>> width $ OpenRange 5 10
5

clipLower :: Ord a => EndPoint a -> Range a -> Maybe (Range a)

Clip the interval from below. I.e. intersect with the interval {l,infty), where { is either open, (, orr closed, [.

clipUpper :: Ord a => EndPoint a -> Range a -> Maybe (Range a)

Clip the interval from above. I.e. intersect with (-infty, u}, where } is either open, ), or closed, ],

midPoint :: Fractional r => Range r -> r

Compute the halfway point between the start and end of a range.

clampTo :: Ord r => Range r -> r -> r

Clamps a value to a range. I.e. if the value lies outside the range we report the closest value "in the range". Note that if an endpoint of the range is open we report that value anyway, so we return a value that is truely inside the range only if that side of the range is closed.

>>> clampTo (ClosedRange 0 10) 20
10
>>> clampTo (ClosedRange 0 10) (-20)
0
>>> clampTo (ClosedRange 0 10) 5
5
>>> clampTo (OpenRange 0 10) 20
10
>>> clampTo (OpenRange 0 10) (-20)
0
>>> clampTo (OpenRange 0 10) 5
5

isValid :: Ord a => Range a -> Bool

Check if the range is valid and nonEmpty, i.e. if the lower endpoint is indeed smaller than the right endpoint. Note that we treat empty open-ranges as invalid as well.

covers :: forall a. Ord a => Range a -> Range a -> Bool

Wether or not the first range completely covers the second one

shiftLeft :: Num r => r -> Range r -> Range r

Shift a range x units to the left

>>> prettyShow $ shiftLeft 10 (ClosedRange 10 20)
"[0,10]"
>>> prettyShow $ shiftLeft 10 (OpenRange 15 25)
"(5,15)"

shiftRight :: Num r => r -> Range r -> Range r

Shifts the range to the right

>>> prettyShow $ shiftRight 10 (ClosedRange 10 20)
"[20,30]"
>>> prettyShow $ shiftRight 10 (OpenRange 15 25)
"(25,35)"