Data.RealNumber.Rational

Description

Documentation

newtype RealNumber (p :: Nat)

Real Numbers represented using Rational numbers. The number type itself is exact in the sense that we can represent any rational number.

The parameter, a natural number, represents the precision (in number of decimals behind the period) with which we display the numbers when printing them (using Show).

If the number cannot be displayed exactly a '~' is printed after the number.

Constructors

RealNumber Rational

Converting to and from RealNumber's

data AsFixed p

Fixed-precision representation of a RealNumber. If there's insufficient precision to accurately represent the RealNumber then the Lossy constructor will be used.

Constructors

Exact !(Fixed p)

Lossy !(Fixed p)

asFixed :: KnownNat p => RealNumber p -> AsFixed (NatPrec p)

Cast RealNumber to a fixed-precision number. Data-loss caused by insufficient precision will be marked by the Lossy constructor.

toFixed :: KnownNat p => RealNumber p -> Fixed (NatPrec p)

Cast RealNumber to a fixed-precision number. Data is silently lost if there's insufficient precision.

fromFixed :: KnownNat p => Fixed (NatPrec p) -> RealNumber p

Cast a fixed-precision number to a RealNumber.

data Nat

(Kind) This is the kind of type-level natural numbers.