Data.UnBounded

Description

Add an unbounded/infintity element to a data type. Essentially, Bottom adds \(-\infty\) (and is pretty much identical to Maybe), whereas Top adds \(\infty\). The Unbounded type adds both.

Documentation

data Top a

`Top a` represents the type a, together with a Top element, i.e. an element that is greater than any other element. We can think of `Top a` being defined as:

>>> data Top a = ValT a | Top

topToMaybe :: Top a -> Maybe a

pattern ValT :: a -> Top a

pattern Top :: Top a

_ValT :: Prism (Top a) (Top b) a b

_Top :: Prism' (Top a) ()

_TopMaybe :: Iso' (Top a) (Maybe a)

Iso between a 'Top a' and a 'Maybe a', interpreting a Top as a Nothing and vice versa. Note that this reverses the ordering of the elements.

>>> ValT 5 ^. _TopMaybe
Just 5
>>> Just 5 ^.re _TopMaybe
ValT 5
>>> Top ^. _TopMaybe
Nothing
>>> Nothing ^.re _TopMaybe
Top

data Bottom a

`Bottom a` represents the type a, together with a Bottom element, i.e. an element that is smaller than any other element. We can think of `Bottom a` being defined as:

>>> data Bottom a = Bottom | ValB a

bottomToMaybe :: Bottom a -> Maybe a

pattern Bottom :: Bottom a

pattern ValB :: a -> Bottom a

_ValB :: Prism (Bottom a) (Bottom b) a b

_Bottom :: Prism' (Bottom a) ()

_BottomMaybe :: Iso' (Bottom a) (Maybe a)

Iso between a 'Bottom a' and a 'Maybe a', interpreting a Bottom as a Nothing and vice versa.

>>> ValB 5 ^. _BottomMaybe
Just 5
>>> Just 5 ^.re _BottomMaybe
ValB 5
>>> Bottom ^. _BottomMaybe
Nothing
>>> Nothing ^.re _BottomMaybe
Bottom

data UnBounded a

`UnBounded a` represents the type a, together with an element MaxInfinity larger than any other element, and an element MinInfinity, smaller than any other element.

Constructors

MinInfinity

Val

Fields

_unUnBounded :: a

MaxInfinity

unUnBounded :: forall a a. Traversal (UnBounded a) (UnBounded a) a a

_MinInfinity :: forall a. Prism' (UnBounded a) ()

_Val :: forall a a. Prism (UnBounded a) (UnBounded a) a a

_MaxInfinity :: forall a. Prism' (UnBounded a) ()

unBoundedToMaybe :: UnBounded a -> Maybe a

Test if an Unbounded is actually bounded.

>>> unBoundedToMaybe (Val 5)
Just 5
>>> unBoundedToMaybe MinInfinity
Nothing
>>> unBoundedToMaybe MaxInfinity
Nothing