HGeometry.Cyclic

Description

Representing Cyclic Sequences

Documentation

newtype Cyclic v a

A cyclic sequence type

Constructors

Cyclic (v a)

toCircularVector :: Cyclic NonEmptyVector a -> CircularVector a

Turn the cyclic vector into a circular Vector

class HasDirectedTraversals v where

Class that models that some type has a cyclic traversal starting from a particular index.

Methods

traverseRightFrom :: Index (v a) -> IndexedTraversal1' (Index (v a)) (v a) a

A rightward-traversal over all elements starting from the given one.

running time : \(O(n)\)

traverseLeftFrom :: Index (v a) -> IndexedTraversal1' (Index (v a)) (v a) a

A rightward-traversal over all elements starting from the given one.

running time : \(O(n)\)

class ShiftedEq t where

Class for types that have an Equality test up to cyclic shifts

Associated Types

type ElemCyclic t

The type of the elements stored in this cyclic container.

Methods

isShiftOf :: Eq (ElemCyclic t) => t -> t -> Bool

Given a and b, test if a is a shifted version of the other.