HGeometry.Ext

Description

A pair-like data type to represent a core type that has extra information as well.

Documentation

data core :+ extra infixr 1

Our Ext type that represents the core datatype core extended with extra information of type extra.

Constructors

core :+ extra infixr 1

_core :: (core :+ extra) -> core

Access the core of an extended value.

_extra :: (core :+ extra) -> extra

Access the extra part of an extended value.

core :: Lens (core :+ extra) (core' :+ extra) core core'

Lens access to the core of an extended value.

extra :: Lens (core :+ extra) (core :+ extra') extra extra'

Lens access to the extra part of an extended value.

ext :: a -> a :+ ()

Tag a value with the unit type.

class AsA t c | t -> c where

A class for types that can behave as a c. Mostly for types t that "extend" a core type c.

Methods

asCore :: t -> c

Get the core from the t.

type (:~) t c = t `AsA` c

infifx shorthand for AsA

pattern AsA :: t :~ c => c -> t

Pattern to get the core.

class AsExt t where

Types that can be decomposed into an Ext

Associated Types

type CoreOf t

type ExtraOf t

Methods

_Ext :: Iso' t (CoreOf t :+ ExtraOf t)

Convert between this type and an Ext