HGeometry.Indexed

Description

Things that have an index.

Documentation

class HasIndex a where

Types that have an index.

Methods

sosIndex :: a -> Index

Get the index of a given a

type Index = Int

A type of Indices

data WithIndex a

Associate a value of type a with an Index.

Constructors

WithIndex !Index a

theValue :: Lens (WithIndex a) (WithIndex b) a b

Lens to manipulate the value of the WithIndex

labelWithIndex :: Traversable t => t a -> t (WithIndex a)

Label each element with its index.

labelWith :: Traversable t => (Index -> a -> b) -> t a -> t b

Label each element with its index using the given labelling function.

labelWith' :: forall t a b. Traversable t => (Index -> a -> b) -> t a -> (t b, Int)

Label each element with its index using the given labelling function. Returns the new collection as well as its size.