HGeometry.Measured.Report

Description

Documentation

newtype Report a

Report is a type to be used for reporting/collecting all results of a query.

The result is internally represented as a binary tree, so that we can guarantee O(1) time and space (<>).

Constructors

MkReport (Maybe (Report1 a))

Bundled Patterns

pattern Report :: [a] -> Report a
Get the results of a Report as a list.

newtype Report1 a

Type to represent reporting elements. The result is represented as a binary tree so that we can support O(1) time and space (<>)

Constructors

Report1 (BinLeafTree (Const () a) a)

newtype ReportList a

Report is a type to be used for reporting/collecting all results of a query. The result is directly represented as a list.

Constructors

ReportList [a]