HGeometry.Sign

Description

Signs of expressions

Documentation

data Sign

The sign of an expression

Constructors

Negative

Positive

flipSign :: Sign -> Sign

Flip Positive = Negative.

signFromTerms :: (Num r, Eq r) => [r] -> Sign

Given the terms, in decreasing order of significance, computes the sign

i.e. expects a list of terms, we base the sign on the sign of the first non-zero term.

pre: the list contains at least one such a term.