HGeometry.Tree.Binary.Static

Description

Several types of Binary trees.

Documentation

data BinLeafTree v a

Binary tree that stores its values (of type a) in the leaves. Internal nodes store something of type v.

Constructors

Leaf !a

Node (BinLeafTree v a) !v (BinLeafTree v a)

node :: (Measured f a, Semigroup (f a)) => BinLeafTree (f a) a -> BinLeafTree (f a) a -> BinLeafTree (f a) a

smart constructor

asBalancedBinLeafTree :: Foldable1 f => f a -> BinLeafTree (Count a) a

Create a balanced tree, i.e. a tree of height \(O(\log n)\) with the elements in the leaves.

\(O(n)\) time.

foldUp :: (b -> v -> b -> b) -> (a -> b) -> BinLeafTree v a -> b

Given a function to combine internal nodes into b's and leafs into b's, traverse the tree bottom up, and combine everything into one b.

foldUpData :: (w -> v -> w -> w) -> (a -> w) -> BinLeafTree v a -> BinLeafTree w a

Traverses the tree bottom up, recomputing the assocated values.

zipExactWith :: (u -> v -> w) -> (a -> b -> c) -> BinLeafTree u a -> BinLeafTree v b -> BinLeafTree w c

Takes two trees, that have the same structure, and uses the provided functions to "zip" them together

toRoseTree :: BinLeafTree v a -> Tree (TreeNode v a)

\( O(n) \) Convert binary tree to a rose tree, aka Tree.

drawTree :: (Show v, Show a) => BinLeafTree v a -> String

2-dimensional ASCII drawing of a tree.

data BinaryTree a

Binary tree in which we store the values of type a in internal nodes.

Constructors

Nil

Internal (BinaryTree a) !a (BinaryTree a)

asBalancedBinTree :: forall f a. Foldable f => f a -> BinaryTree a

Create a balanced binary tree.

running time: \(O(n)\)

access :: BinaryTree a -> Maybe a

Get the element stored at the root, if it exists

foldBinaryUp :: b -> (a -> b -> b -> b) -> BinaryTree a -> BinaryTree (a, b)

Fold function for folding over a binary tree.

toRoseTree' :: BinaryTree a -> Maybe (Tree a)

Convert a BinaryTree into a RoseTree

drawTree' :: Show a => BinaryTree a -> String

Draw a binary tree.