Data.OrdSeq

Documentation

data Key a

Constructors

NoKey

Key

Fields

getKey :: !a

liftCmp :: (a -> a -> Ordering) -> Key a -> Key a -> Ordering

newtype Elem a

Constructors

Elem

Fields

getElem :: a

newtype OrdSeq a

Constructors

OrdSeq

Fields

_asFingerTree :: FingerTree (Key a) (Elem a)

type Compare a = a -> a -> Ordering

insertBy :: Compare a -> a -> OrdSeq a -> OrdSeq a

Insert into a monotone OrdSeq.

pre: the comparator maintains monotonicity

\(O(\log n)\)

insert :: Ord a => a -> OrdSeq a -> OrdSeq a

Insert into a sorted OrdSeq

\(O(\log n)\)

deleteAllBy :: Compare a -> a -> OrdSeq a -> OrdSeq a

splitBy :: Compare a -> a -> OrdSeq a -> (OrdSeq a, OrdSeq a, OrdSeq a)

\(O(\log n)\)

splitOn :: Ord b => (a -> b) -> b -> OrdSeq a -> (OrdSeq a, OrdSeq a, OrdSeq a)

Given a monotonic function f that maps a to b, split the sequence s depending on the b values. I.e. the result (l,m,r) is such that * all (< x) . fmap f $ l * all (== x) . fmap f $ m * all (> x) . fmap f $ r

>>> splitOn id 3 $ fromAscList' [1..5]
(OrdSeq {_asFingerTree = fromList [Elem 1,Elem 2]},OrdSeq {_asFingerTree = fromList [Elem 3]},OrdSeq {_asFingerTree = fromList [Elem 4,Elem 5]})
>>> splitOn fst 2 $ fromAscList' [(0,"-"),(1,"A"),(2,"B"),(2,"C"),(3,"D"),(4,"E")]
(OrdSeq {_asFingerTree = fromList [Elem (0,"-"),Elem (1,"A")]},OrdSeq {_asFingerTree = fromList [Elem (2,"B"),Elem (2,"C")]},OrdSeq {_asFingerTree = fromList [Elem (3,"D"),Elem (4,"E")]})

\(O(\log n)\)

splitMonotonic :: (a -> Bool) -> OrdSeq a -> (OrdSeq a, OrdSeq a)

Given a monotonic predicate p, splits the sequence s into two sequences (as,bs) such that all (not p) as and all p bs

\(O(\log n)\)

deleteAll :: Ord a => a -> OrdSeq a -> OrdSeq a

fromListBy :: Compare a -> [a] -> OrdSeq a

inserts all eleements in order \(O(n\log n)\)

fromListByOrd :: Ord a => [a] -> OrdSeq a

inserts all eleements in order \(O(n\log n)\)

fromAscList' :: [a] -> OrdSeq a

O(n)

lookupBy :: Compare a -> a -> OrdSeq a -> Maybe a

\(O(\log n)\)

memberBy :: Compare a -> a -> OrdSeq a -> Bool

mapMonotonic :: (a -> b) -> OrdSeq a -> OrdSeq b

Fmap, assumes the order does not change O(n)

viewl :: OrdSeq a -> ViewL OrdSeq a

Gets the first element from the sequence \(O(1)\)

viewr :: OrdSeq a -> ViewR OrdSeq a

minView :: OrdSeq a -> Maybe (a, OrdSeq a)

lookupMin :: OrdSeq a -> Maybe a

maxView :: OrdSeq a -> Maybe (a, OrdSeq a)

lookupMax :: OrdSeq a -> Maybe a