Data.UnBounded

Description

Add an unbounded/infintity element to a data type. Essentially, Bottom adds \(-\infty\) (and is pretty much identical to Maybe), whereas Top adds \(\infty\). The UnBounded type adds both.

Documentation

data Top a where

Top a represents the type a, together with a Top element, i.e. an element that is greater than any other element. We can think of `Top a` being defined as:

>>> data Top a = ValT a | Top

Bundled Patterns

pattern ValT :: a -> Top a

pattern Top :: Top a

topToMaybe :: Top a -> Maybe a

Top a values are isomorphing to Maybe a values.

_ValT :: Prism (Top a) (Top b) a b

ValT prism. Can be used to access the non-bottom element if it exists:

>>> ValT True & _ValT %~ not
ValT False

>>> Top & _ValT %~ not
Top

_Top :: Prism' (Top a) ()

Top prism.

_TopMaybe :: Iso' (Top a) (Maybe a)

Iso between a Top a and a Maybe a, interpreting a Top as a Nothing and vice versa. Note that this reverses the ordering of the elements.

>>> ValT 5 ^. _TopMaybe
Just 5
>>> Just 5 ^.re _TopMaybe
ValT 5
>>> Top ^. _TopMaybe
Nothing
>>> Nothing ^.re _TopMaybe
Top

data Bottom a where

`Bottom a` represents the type a, together with a Bottom element, i.e. an element that is smaller than any other element. We can think of `Bottom a` being defined as:

>>> data Bottom a = Bottom | ValB a

Bundled Patterns

pattern Bottom :: Bottom a

pattern ValB :: a -> Bottom a

bottomToMaybe :: Bottom a -> Maybe a

`Bottom a` values are isomorphing to `Maybe a` values.

_ValB :: Prism (Bottom a) (Bottom b) a b

ValB prism. Can be used to access the non-bottom element if it exists:

>>> ValB True & _ValB %~ not
ValB False

>>> Bottom & _ValB %~ not
Bottom

_Bottom :: Prism' (Bottom a) ()

Bottom prism.

_BottomMaybe :: Iso' (Bottom a) (Maybe a)

Iso between a 'Bottom a' and a 'Maybe a', interpreting a Bottom as a Nothing and vice versa.

>>> ValB 5 ^. _BottomMaybe
Just 5
>>> Just 5 ^.re _BottomMaybe
ValB 5
>>> Bottom ^. _BottomMaybe
Nothing
>>> Nothing ^.re _BottomMaybe
Bottom

data UnBounded a

`UnBounded a` represents the type a, together with an element MaxInfinity larger than any other element, and an element MinInfinity, smaller than any other element.

Constructors

MinInfinity

Val

Fields

_unUnBounded :: a

MaxInfinity

_Val :: Prism (UnBounded a) (UnBounded b) a b

Prism to access unbounded value if it exists.

>>> Val True ^? _Val
Just True

>>> MinInfinity ^? _Val :: Maybe Bool
Nothing

>>> Val True & _Val %~ not
Val False

>>> MaxInfinity & _Val %~ not
MaxInfinity

unBoundedToMaybe :: UnBounded a -> Maybe a

Test if an Unbounded is actually bounded.

>>> unBoundedToMaybe (Val 5)
Just 5
>>> unBoundedToMaybe MinInfinity
Nothing
>>> unBoundedToMaybe MaxInfinity
Nothing